% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make.R
\name{covars_make_baselines}
\alias{covars_make_baselines}
\title{Compute baseline covariates}
\usage{
covars_make_baselines(x, reference_year = NULL)
}
\arguments{
\item{x}{a \link[quanteda:corpus]{quanteda::corpus} object or character vector}

\item{reference_year}{optional reference year for temporal normalization}
}
\value{
a list of baseline statistics
}
\description{
Computes mean and standard deviation of covariates for normalization.
}
