% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophistication2-package.R
\docType{package}
\name{sophistication2-package}
\alias{sophistication2-package}
\alias{sophistication2}
\title{sophistication2: Measuring Text Sophistication}
\description{
Provides functions for measuring the sophistication of political texts based on
readability and linguistic features. Includes tools for creating text snippets,
generating comparison pairs, computing readability scores with bootstrap estimates,
and predicting text sophistication.
}
\section{Main Functions}{

\itemize{
\item \code{\link{snippets_make}} - Create text snippets from corpus
\item \code{\link{snippets_clean}} - Clean problematic snippets
\item \code{\link{pairs_regular_make}} - Generate snippet pairs
\item \code{\link{pairs_gold_make}} - Create gold standard pairs
\item \code{\link{bootstrap_readability}} - Bootstrap readability statistics
\item \code{\link{covars_make}} - Compute text covariates
\item \code{\link{covars_make_pos}} - Compute POS features (using udpipe)
\item \code{\link{predict_readability}} - Predict readability scores
}
}

\section{Language Models}{

The \code{\link{covars_make_pos}} function uses udpipe for part-of-speech tagging.
On first use, it automatically downloads the English language model (~17MB).
This is a pure R solution - no Python or external dependencies required.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbenoit/sophistication}
  \item Report bugs at \url{https://github.com/kbenoit/sophistication/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kevin Munger \email{kevin.munger@gmail.com}
  \item Arthur Spirling \email{as9934@nyu.edu}
}

Other contributors:
\itemize{
  \item Arthur Spirling Updater \email{maintainer@example.com} (Updated for modern quanteda) [contributor]
}

}
\keyword{internal}
